+++ CHANGELOG.md
@@ -6,10 +6,12 @@ It is based on Tidepool Blip 1.27.
 ### Added
 - YLP-3613 Memos on DBL settings in Devices view
 - YLP-3617 As a user, I can see the devices changes history
+- YLP-3785 Be able to know the effective DBLG1 or DBLG2 settings for the displayed day in the daily view
 
 ### Changed
 - YLP-2823 Use , instead of . for languages other than English
 - YLP-3618 Meal without bolus management in Daily view
+- YLP-3634 Add tooltips for Rescue carbs and Manual bolus metrics
 - YLP-3815 Display update timestamp in yourloops
 - YLP-3830 Use a generic CGM icon for warmup events